pagination

+++ src/main/java/com/springboot/App/DataAccessLayer/routeController/pagedResponse.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class pagedResponse<T> {
     public List<T> items = new ArrayList<T>();
 
-    public int totalItems = 0;
+    public long totalItems = 0;
 
     public int itemsPerPage = 0;
